requirements.txt
click==7.1.2
Flask==1.1.1
Flask-SQLAlchemy==2.4
flask-restx==0.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==0.16.1
